third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-override-custom-property.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#e {
    width: 100px;

    --foo: {
        --before: green;
        --after: 0px;
    };

    --before: red;
    @apply --foo;
    --after: 200px;

    height: var(--after);
    background-color: var(--before);
}
</style>
<div id=e>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(e).width, "100px");
    assert_equals(getComputedStyle(e).height, "200px");
}, "@apply can override custom properties");
</script>
