third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-override-custom-property-expected.txt

This is a testharness.js-based test.
FAIL @apply can override custom properties assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
